+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -114,7 +114,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-02-10T22:25:10.468Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-02-11T07:51:04.718Z")
     r"""The date and time the resource was created"""