+++ packages/orq-rc/RELEASES.md
@@ -329,3 +329,13 @@ Based on:
 - [python v3.1.0-rc.55] packages/orq-rc
 ### Releases
 - [PyPI v3.1.0-rc.55] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.55 - packages/orq-rc
+
+## 2025-02-11 07:52:23
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.488.0 (2.506.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.1.0-rc.56] packages/orq-rc
+### Releases
+- [PyPI v3.1.0-rc.56] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.56 - packages/orq-rc